Initial commit
This is a copy of a project located in https://github.com/paepcke/ElephantCallAI. This repo exists only for turning in the source code of this assignment. There will be references to utilities outside the 'gun_data' directory of the aforementioned repo, so this repo will not work outside of that context. It can, however, be read by a TA :)